make base_gpt.py and move shared methods over
